\name{print.GAPerm}
\alias{print.GAPerm}
\alias{print.summaryGAPerm}
\title{Print GA results}
\usage{
  \method{print}{GAPerm} (x, ...)

  \method{print}{summaryGAPerm} (x, ...)
}
\arguments{
  \item{x}{An object of class \code{GAPerm} or
  \code{summaryGAPerm}}

  \item{...}{Other parameters (will be ignored).}
}
\description{
  Prints the GA results.
}

