\name{print.GAReal}
\alias{print.GAReal}
\alias{print.summaryGAReal}
\title{Print GA results}
\usage{
  \method{print}{GAReal} (x, ...)

  \method{print}{summaryGAReal} (x, ...)
}
\arguments{
  \item{x}{An object of class \code{GAReal} or
  \code{summaryGAReal}}

  \item{...}{Other parameters (will be ignored).}
}
\description{
  Prints the GA results.
}

