\name{summary.GAPerm}
\alias{summary.GAPerm}
\title{Genetic Algorithm summary}
\usage{
  \method{summary}{GAPerm} (object, ...)
}
\arguments{
  \item{object}{An object of class \code{GAPerm}.}

  \item{...}{Other parameters (will be ignored).}
}
\value{
  An object of class \code{summaryGAPerm}, which is a list
  that can be inspected or printed on-screen.
}
\description{
  A simple summary of the GA results.
}

