\name{summary.GAReal}
\alias{summary.GAReal}
\title{Genetic Algorithm summary}
\usage{
  \method{summary}{GAReal} (object, ...)
}
\arguments{
  \item{object}{An object of class \code{GAReal}.}

  \item{...}{Other parameters (will be ignored).}
}
\value{
  An object of class \code{summaryGAReal}, which is a list
  that can be inspected or printed on-screen.
}
\description{
  A simple summary of the GA results.
}

